+++ R/zzz.R
@@ -20,7 +20,7 @@
         "computePolicy", "decrypt_envelope", "epsilon",
         "funs", "n", "nOrders",  "path_T1", "path_T3",
         "rewardseq.OFF", "rewardseq.ON",
-        "totreward", "trstate", "MarketType",
+        "totreward", "trstate", "MarketType","predict_CMSUM", "LABEL_CMSUM",
 
 
         # sample function names
@@ -30,7 +30,8 @@
         "h2o.deeplearning", "h2o.loadModel", "h2o.predict", "h2o.saveModel", "head", "tail",
         "if_else", "lag", "mutate_all", "mutate_if", "slice", "str_detect",
         "predict", "read_rds", "summarise", "write.csv", "write_rds", "ymd_hms",
-        "filter_all", "full_join", "bind_rows",  "as.h2o", "bind_cols", "testdir"
+        "filter_all", "full_join", "bind_rows",  "as.h2o", "bind_cols", "testdir",
+        "h2o.performance"
       )
     )