+++ NEWS.md
@@ -11,6 +11,7 @@
 * building a functionality to create separate models for each asset (adding functions with prefix aml_xxx_xxx.R)
 * build more simple functions to collect data, build model, score data to predict
 * added tests for functions collect data and test model
+* function aml_make_model() will be finding the best neural network structure automatically using 6 random structures
 
 # lazytrade 0.3.4